refactor : 패키지 분리 고도화, 레거시 코드 삭제
