+++ src/main/java/lotto/model/lotto/Lotto.java
@@ -1,19 +1,18 @@
-package lotto.model;
+package lotto.model.lotto;
 
 import java.util.Collections;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
 
-public class WinningNumber {
+public class Lotto {
     private final List<Integer> numbers;
 
-    public WinningNumber(List<Integer> numbers) {
+    public Lotto(List<Integer> numbers) {
         validate(numbers);
         this.numbers = numbers;
     }
 
-    //Lotto와 검증로직 겹침, 전환로직과 분리 필요
     private void validate(List<Integer> numbers) {
         if (numbers.size() != 6) {
             throw new IllegalArgumentException("[ERROR] 로또 번호는 6개여야 합니다.");
@@ -37,11 +36,18 @@ public class WinningNumber {
         return !uniqueNumbers.add(number);
     }
 
-    public boolean contains(int number) {
-        if (numbers.contains(number)) {
-            return true;
+    public int checkMatchingAmountWith(List<Integer> numbers) {
+        int matchingAmount = 0;
+        for (int number : numbers) {
+            if (contains(number)) {
+                matchingAmount++;
+            }
         }
-        return false;
+        return matchingAmount;
+    }
+
+    public boolean contains(int number) {
+        return numbers.contains(number);
     }
 
     public List<Integer> getNumbers() {
