+++ src/test/java/lotto/model/BonusNumberTest.java
@@ -2,6 +2,7 @@ package lotto.model;
 
 import static org.assertj.core.api.Assertions.assertThatIllegalArgumentException;
 
+import lotto.model.winningNumber.BonusNumber;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
