+++ src/test/java/lotto/model/LottoTest.java
@@ -3,6 +3,7 @@ package lotto.model;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.stream.Stream;
+import lotto.model.lotto.Lotto;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
@@ -36,17 +37,6 @@ class LottoTest {
 
         assertThat(lotto.getNumbers()).contains(testNumbers.get(lottoNumbersIndex));
     }
-
-//    @ParameterizedTest
-//    @ValueSource(ints = {0, 1, 2, 3, 4})
-//    void 숫자를_뽑을_때_오름차순으로_정렬한다(int lottoNumbersIndex) {
-//        List<Integer> randomNumbers = RandomNumberPicker.pickAscendingNumbers();
-//
-//        Lotto lotto = new Lotto(randomNumbers);
-//
-//        assertThat(lotto.getNumbers().get(lottoNumbersIndex))
-//                .isLessThan(lotto.getNumbers().get(lottoNumbersIndex + 1));
-//    }
 
     @ParameterizedTest
     @MethodSource("lottoWithMatchingNumberProvider")
