+++ src/test/java/lotto/model/NumberGeneratorTest.java
@@ -6,6 +6,8 @@ import static org.assertj.core.api.Assertions.assertThatIllegalArgumentException
 
 import java.util.ArrayList;
 import java.util.Arrays;
+import lotto.model.winningNumber.NumberGenerator;
+import lotto.model.winningNumber.WinningNumber;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
