+++ src/test/java/lotto/model/WinningRankTest.java
@@ -3,6 +3,7 @@ package lotto.model;
 import java.util.Arrays;
 import java.util.List;
 import java.util.stream.Stream;
+import lotto.model.winningResult.WinningRank;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.MethodSource;
