+++ src/test/java/lotto/model/WinningResultsTest.java
@@ -2,6 +2,8 @@ package lotto.model;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import lotto.model.winningResult.WinningRank;
+import lotto.model.winningResult.WinningResults;
 import org.junit.jupiter.api.Test;
 
 public class WinningResultsTest {
