+++ src/test/java/lotto/service/LottoMachineTest.java
@@ -5,13 +5,13 @@ import static org.assertj.core.api.Assertions.assertThat;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
-import lotto.model.BonusNumber;
-import lotto.model.Lotto;
-import lotto.model.PurchaseAmount;
-import lotto.model.Lottos;
-import lotto.model.WinningNumber;
-import lotto.model.WinningRank;
-import lotto.model.WinningResults;
+import lotto.model.winningNumber.BonusNumber;
+import lotto.model.lotto.Lotto;
+import lotto.model.lotto.PurchaseAmount;
+import lotto.model.lotto.Lottos;
+import lotto.model.winningNumber.WinningNumber;
+import lotto.model.winningResult.WinningRank;
+import lotto.model.winningResult.WinningResults;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
